+++ zellij-server/src/tab/mod.rs
@@ -157,6 +157,7 @@ pub(crate) struct Tab {
     viewport: Rc<RefCell<Viewport>>, // includes all non-UI panes
     display_area: Rc<RefCell<Size>>, // includes all panes (including eg. the status bar and tab bar in the default layout)
     character_cell_size: Rc<RefCell<Option<SizeInPixels>>>,
+    stacked_resize: Rc<RefCell<bool>>,
     sixel_image_store: Rc<RefCell<SixelImageStore>>,
     os_api: Box<dyn ServerOsApi>,
     pub senders: ThreadSenders,
@@ -187,7 +188,8 @@ pub(crate) struct Tab {
     pending_instructions: Vec<BufferedTabInstruction>, // instructions that came while the tab was
     // pending and need to be re-applied
     swap_layouts: SwapLayouts,
-    default_shell: Option<PathBuf>,
+    default_shell: PathBuf,
+    default_editor: Option<PathBuf>,
     debug: bool,
     arrow_fonts: bool,
     styled_underlines: bool,
@@ -228,6 +230,7 @@ pub trait Pane {
         _key_with_modifier: &Option<KeyWithModifier>,
         _raw_input_bytes: Vec<u8>,
         _raw_input_bytes_are_kitty: bool,
+        _client_id: Option<ClientId>,
     ) -> Option<AdjustedInput> {
         None
     }
@@ -564,6 +567,7 @@ impl Tab {
         name: String,
         display_area: Size,
         character_cell_size: Rc<RefCell<Option<SizeInPixels>>>,
+        stacked_resize: Rc<RefCell<bool>>,
         sixel_image_store: Rc<RefCell<SixelImageStore>>,
         os_api: Box<dyn ServerOsApi>,
         senders: ThreadSenders,
@@ -579,11 +583,12 @@ impl Tab {
         terminal_emulator_colors: Rc<RefCell<Palette>>,
         terminal_emulator_color_codes: Rc<RefCell<HashMap<usize, String>>>,
         swap_layouts: (Vec<SwapTiledLayout>, Vec<SwapFloatingLayout>),
-        default_shell: Option<PathBuf>,
+        default_shell: PathBuf,
         debug: bool,
         arrow_fonts: bool,
         styled_underlines: bool,
         explicitly_disable_kitty_keyboard_protocol: bool,
+        default_editor: Option<PathBuf>,
     ) -> Self {
         let name = if name.is_empty() {
             format!("Tab #{}", index + 1)
@@ -608,6 +613,7 @@ impl Tab {
             connected_clients_in_app.clone(),
             mode_info.clone(),
             character_cell_size.clone(),
+            stacked_resize.clone(),
             session_is_mirrored,
             draw_pane_frames,
             default_mode_info.clone(),
@@ -647,6 +653,7 @@ impl Tab {
             viewport,
             display_area,
             character_cell_size,
+            stacked_resize,
             sixel_image_store,
             synchronize_is_active: false,
             os_api,
@@ -677,6 +684,7 @@ impl Tab {
             arrow_fonts,
             styled_underlines,
             explicitly_disable_kitty_keyboard_protocol,
+            default_editor,
         }
     }
 
@@ -881,8 +889,13 @@ impl Tab {
         let mode_infos = self.mode_info.borrow();
         let mut plugin_updates = vec![];
         for client_id in self.connected_clients.borrow().iter() {
-            let mode_info = mode_infos.get(client_id).unwrap_or(&self.default_mode_info);
-            plugin_updates.push((None, Some(*client_id), Event::ModeUpdate(mode_info.clone())));
+            let mut mode_info = mode_infos
+                .get(client_id)
+                .unwrap_or(&self.default_mode_info)
+                .clone();
+            mode_info.shell = Some(self.default_shell.clone());
+            mode_info.editor = self.default_editor.clone();
+            plugin_updates.push((None, Some(*client_id), Event::ModeUpdate(mode_info)));
         }
         self.senders
             .send_to_plugin(PluginInstruction::Update(plugin_updates))
@@ -924,6 +937,8 @@ impl Tab {
             };
             self.tiled_panes
                 .focus_pane_if_client_not_focused(focus_pane_id, client_id);
+            self.floating_panes
+                .focus_first_pane_if_client_not_focused(client_id);
             self.connected_clients.borrow_mut().insert(client_id);
             self.mode_info.borrow_mut().insert(
                 client_id,
@@ -1011,7 +1026,7 @@ impl Tab {
             }
             if let Some(embedded_pane_to_float) = self.extract_pane(focused_pane_id, true) {
                 self.show_floating_panes();
-                self.add_floating_pane(embedded_pane_to_float, focused_pane_id, None)?;
+                self.add_floating_pane(embedded_pane_to_float, focused_pane_id, None, true)?;
             }
         }
         Ok(())
@@ -1043,7 +1058,7 @@ impl Tab {
                 return Ok(());
             }
             if let Some(embedded_pane_to_float) = self.extract_pane(pane_id, true) {
-                self.add_floating_pane(embedded_pane_to_float, pane_id, None)?;
+                self.add_floating_pane(embedded_pane_to_float, pane_id, None, true)?;
             }
         }
         Ok(())
@@ -1188,7 +1203,7 @@ impl Tab {
                 .insert(pid, (is_scrollback_editor, new_pane));
             Ok(())
         } else if self.floating_panes.panes_are_visible() {
-            self.add_floating_pane(new_pane, pid, floating_pane_coordinates)
+            self.add_floating_pane(new_pane, pid, floating_pane_coordinates, true)
         } else {
             self.add_tiled_pane(new_pane, pid, client_id)
         }
@@ -1870,6 +1885,7 @@ impl Tab {
                     key_with_modifier,
                     raw_input_bytes,
                     raw_input_bytes_are_kitty,
+                    client_id,
                 ) {
                     Some(AdjustedInput::WriteBytesToTerminal(adjusted_input)) => {
                         self.senders
@@ -1898,7 +1914,7 @@ impl Tab {
                         self.senders
                             .send_to_pty(PtyInstruction::DropToShellInPane {
                                 pane_id: PaneId::Terminal(active_terminal_id),
-                                shell: self.default_shell.clone(),
+                                shell: Some(self.default_shell.clone()),
                                 working_dir,
                             })
                             .with_context(err_context)?;
@@ -1912,6 +1928,7 @@ impl Tab {
                 key_with_modifier,
                 raw_input_bytes,
                 raw_input_bytes_are_kitty,
+                client_id,
             ) {
                 Some(AdjustedInput::WriteKeyToPlugin(key_with_modifier)) => {
                     self.senders
@@ -4160,7 +4177,7 @@ impl Tab {
                     pane.1.set_selectable(true);
                     if should_float {
                         self.show_floating_panes();
-                        self.add_floating_pane(pane.1, pane_id, None)
+                        self.add_floating_pane(pane.1, pane_id, None, true)
                     } else {
                         self.hide_floating_panes();
                         self.add_tiled_pane(pane.1, pane_id, Some(client_id))
@@ -4173,7 +4190,8 @@ impl Tab {
         match self.suppressed_panes.remove(&pane_id) {
             Some(pane) => {
                 self.show_floating_panes();
-                self.add_floating_pane(pane.1, pane_id, None).non_fatal();
+                self.add_floating_pane(pane.1, pane_id, None, true)
+                    .non_fatal();
                 self.floating_panes.focus_pane_for_all_clients(pane_id);
             },
             None => {
@@ -4213,6 +4231,7 @@ impl Tab {
         mut pane: Box<dyn Pane>,
         pane_id: PaneId,
         floating_pane_coordinates: Option<FloatingPaneCoordinates>,
+        should_focus_new_pane: bool,
     ) -> Result<()> {
         let err_context = || format!("failed to add floating pane");
         if let Some(mut new_pane_geom) = self.floating_panes.find_room_for_new_pane() {
@@ -4230,7 +4249,9 @@ impl Tab {
             resize_pty!(pane, self.os_api, self.senders, self.character_cell_size)
                 .with_context(err_context)?;
             self.floating_panes.add_pane(pane_id, pane);
-            self.floating_panes.focus_pane_for_all_clients(pane_id);
+            if should_focus_new_pane {
+                self.floating_panes.focus_pane_for_all_clients(pane_id);
+            }
         }
         if self.auto_layout && !self.swap_layouts.is_floating_damaged() {
             // only do this if we're already in this layout, otherwise it might be
@@ -4256,9 +4277,10 @@ impl Tab {
             if should_auto_layout {
                 // no need to relayout here, we'll do it when reapplying the swap layout
                 // below
-                self.tiled_panes.insert_pane_without_relayout(pane_id, pane);
+                self.tiled_panes
+                    .insert_pane_without_relayout(pane_id, pane, client_id);
             } else {
-                self.tiled_panes.insert_pane(pane_id, pane);
+                self.tiled_panes.insert_pane(pane_id, pane, client_id);
             }
             self.set_should_clear_display_before_rendering();
             if let Some(client_id) = client_id {
@@ -4346,7 +4368,10 @@ impl Tab {
                 self.set_force_render(); // we force render here to make sure the panes under the floating pane render and don't leave "garbage" in case of a decrease
             }
         } else if self.tiled_panes.panes_contain(&pane_id) {
-            match self.tiled_panes.resize_pane_with_id(strategy, pane_id) {
+            match self
+                .tiled_panes
+                .resize_pane_with_id(strategy, pane_id, None)
+            {
                 Ok(_) => {},
                 Err(err) => match err.downcast_ref::<ZellijError>() {
                     Some(ZellijError::CantResizeFixedPanes { pane_ids }) => {
@@ -4405,8 +4430,15 @@ impl Tab {
             pane.update_arrow_fonts(should_support_arrow_fonts);
         }
     }
-    pub fn update_default_shell(&mut self, default_shell: Option<PathBuf>) {
-        self.default_shell = default_shell;
+    pub fn update_default_shell(&mut self, mut default_shell: Option<PathBuf>) {
+        if let Some(default_shell) = default_shell.take() {
+            self.default_shell = default_shell;
+        }
+    }
+    pub fn update_default_editor(&mut self, mut default_editor: Option<PathBuf>) {
+        if let Some(default_editor) = default_editor.take() {
+            self.default_editor = Some(default_editor);
+        }
     }
     pub fn update_copy_options(&mut self, copy_options: &CopyOptions) {
         self.clipboard_provider = match &copy_options.command {
@@ -4486,6 +4518,35 @@ impl Tab {
         } else if self.tiled_panes.pane_id_is_focused(&root_pane_id) {
             self.tiled_panes.expand_pane_in_stack(root_pane_id);
         }
+    }
+    pub fn change_floating_pane_coordinates(
+        &mut self,
+        pane_id: &PaneId,
+        floating_pane_coordinates: FloatingPaneCoordinates,
+    ) -> Result<()> {
+        if !self.floating_panes.panes_contain(pane_id) {
+            // if these panes are not floating, we make them floating (assuming doing so wouldn't
+            // be removing the last selectable tiled pane in the tab, which would close it)
+            if (self.tiled_panes.panes_contain(&pane_id)
+                && self.get_selectable_tiled_panes().count() <= 1)
+                || self.suppressed_panes.contains_key(pane_id)
+            {
+                if let Some(pane) = self.extract_pane(*pane_id, true) {
+                    self.add_floating_pane(pane, *pane_id, None, false)?;
+                }
+            }
+        }
+        self.floating_panes
+            .change_pane_coordinates(*pane_id, floating_pane_coordinates)?;
+        self.set_force_render();
+        self.swap_layouts.set_is_floating_damaged();
+        Ok(())
+    }
+    pub fn get_viewport(&self) -> Viewport {
+        self.viewport.borrow().clone()
+    }
+    pub fn get_display_area(&self) -> Size {
+        self.display_area.borrow().clone()
     }
     fn new_scrollback_editor_pane(&self, pid: u32) -> TerminalPane {
         let next_terminal_position = self.get_next_terminal_position();