+++ CHANGELOG.md
@@ -20,6 +20,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * feat(terminal): mouse AnyEvent tracking (https://github.com/zellij-org/zellij/pull/3538)
 * fix(terminal): support setting kitty keyboard protocol with `CSI=` (https://github.com/zellij-org/zellij/pull/3942)
 * fix(floating-panes): handle various errors (https://github.com/zellij-org/zellij/pull/3944)
+* chore(rust): Update Rust toolchain to 1.84.0 (https://github.com/zellij-org/zellij/pull/3945)
+* feat(ux): stacked-resize (https://github.com/zellij-org/zellij/pull/3957)
+* feat(plugins): API to change the position and size of floating pane, as well as know the viewport size (https://github.com/zellij-org/zellij/pull/3958 and https://github.com/zellij-org/zellij/pull/3972)
+* feat(plugins): add `PastedText` event (https://github.com/zellij-org/zellij/pull/3962)
+* feat(plugins): APIs to open panes near plugin (https://github.com/zellij-org/zellij/pull/3966)
+* feat(plugins): send information about $SHELL and $EDITOR (https://github.com/zellij-org/zellij/pull/3971)
 
 ## [0.41.2] - 2024-11-19
 * fix(input): keypresses not being identified properly with kitty keyboard protocol in some terminals (https://github.com/zellij-org/zellij/pull/3725)